Added Game Class and built delay function into frame loop
