+++ lib/Player.js
@@ -1,5 +1,4 @@
 const GamePiece = require('./GamePiece.js');
-//const Projectile = require('./Projectile.js')
 
 module.exports = class Player extends GamePiece {
   constructor() {
@@ -59,8 +58,4 @@ module.exports = class Player extends GamePiece {
       this.isOnBoard = true;
     }
   }
-  // fire(x, y, h, w, xd, yd, direction, speed) {
-    // var shot = new Projectile(x,y,h,w,xd,yd,direction,speed)
-    //.push(projectile.bullets);
-  // }
 };
